\name{setTabDefault}
\alias{setTabDefault}
\title{Setting Summary Defaults for Extending mtable()
}
\description{This function resets the summary template for mtable() -- as opposed to using the setSummaryTemplate() command directly. It is called directly from within each getSummary method in tonymisc.
}
\usage{
setTabDefault()
}
\author{J. Anthony Cookson
}


\seealso{ \code{\link{mtable}}  \code{\link{gls}}  \code{\link{ivreg}}
}
