\name{tony.vars}
\alias{tony.vars}
\title{ An alternative all.vars() command
}
\description{ This is a utility command used by iv() that provides a slightly easier-to-use version of all.vars() when the formula statement contains transformations using I().
}
\usage{
tony.vars(x)
}
\arguments{
  \item{x}{ The only argument is a formula statement.
}
}
\value{ Returns a list of variable names.
}
\author{ J. Anthony Cookson
}
