\name{tony.vars2}
\alias{tony.vars2}
\title{ Another alternative all.vars() command
}
\description{ This is a utility command used by iv() that provides a slightly easier-to-use version of all.vars() when the formula statement contains transformations using I().  It also makes it possible to specify/extract multiple LHS variables in the first stage.
}
\usage{
tony.vars2(x)
}
\arguments{
  \item{x}{ The only argument is a formula statement.
}
}
\value{ Returns a list of variable names.
}

\author{ J. Anthony Cookson
}
