\name{tonymisc-package}
\alias{tonymisc-package}
\alias{tonymisc}
\docType{package}
\title{Functions and Extensions for Econometrics Output 
}
\description{This package contains a list of commands that make summarizing regression output simpler.  Much of this package is an extension of the mtable() command in the memisc library, but there are a variety of other methods -- specifically, those for reporting robust standard errors and IV regression output -- that are available in this package.  
}
\details{
\tabular{ll}{
Package: \tab tonymisc\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2011-05-15\cr
License: \tab GPL-2\cr
}

}
\author{
Maintainer: J. Anthony Cookson <tony.cookson@gmail.com>

}

\keyword{ package }
\seealso{
\code{\link[AER]{ivreg}} \code{\link[memisc]{mtable}}
}